demo/python/demo/requirements.txt
pycryptodome==3.20.0
requests==2.32.2
